Poprawki TableView oraz interfejsu graficznego

+++ src/sample/Controllers/zarzadzajUzytkownikamiAdminController.java
@@ -43,7 +43,7 @@ public class zarzadzajUzytkownikamiAdminController implements Initializable  {
     @FXML
     private TableColumn<ModelTable, String> col_miejscowosc;
     @FXML
-    private TableColumn<ModelTable, Double> col_pesel;
+    private TableColumn<ModelTable, Long> col_pesel;
 
     ObservableList<ModelTable> oblist = FXCollections.observableArrayList();
 
@@ -68,7 +68,7 @@ public class zarzadzajUzytkownikamiAdminController implements Initializable  {
             ResultSet rs = con.createStatement().executeQuery("select * from klient");
 
             while (rs.next()){
-                oblist.add(new ModelTable( rs.getString(4), rs.getString(5),rs.getString(6), rs.getString(7),""+ rs.getDouble(8)));
+                oblist.add(new ModelTable( rs.getString(4), rs.getString(5),rs.getString(6), rs.getString(7),""+ rs.getLong(8)));
             }
